\name{template}
\alias{template}
\title{Function Title}
\description{Description of the function}
\usage{

%% Provide an example of how the function works

template(x)
}

\arguments{

%% Each argument of the function needs details. 
%% Please provide object type (character, numeric, list, etc)
%% Describe what needs to be in the argument in plain english

  \item{x}{
}
}

\value{

%%  ~Describe the value returned

%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}



}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or standard data sets, see data().

## The function is currently defined as
function (x) 
{
  }
}

